@@ -5,7 +5,6 @@ from .provider import Infura
 NETWORKS = {
     "ethereum": [
         "mainnet",
-        "goerli",
         "sepolia",
     ],
     "arbitrum": [
@@ -18,13 +17,15 @@ NETWORKS = {
     ],
     "polygon": [
         "mainnet",
-        "mumbai",
         "amoy",
+    ],
+    "blast": [
+        "mainnet",
+        "sepolia",
     ],
     # TODO: Comment out after ape-linea supports 0.7
     # "linea": [
     #     "mainnet",
-    #     "goerli",
     # ],
 }
 
