@@ -19,9 +19,11 @@ def networks():
     return ape.networks
 
 
-@pytest.fixture(params=[(name, net) for name, values in NETWORKS.items() for net in values])
+# NOTE: Using a `str` as param for better pytest test-case name generation.
+@pytest.fixture(params=[f"{e}:{n}" for e, values in NETWORKS.items() for n in values])
 def provider(networks, request):
-    ecosystem_cls = networks.get_ecosystem(request.param[0])
-    network_cls = ecosystem_cls.get_network(request.param[1])
+    ecosystem, network = request.param.split(":")
+    ecosystem_cls = networks.get_ecosystem(ecosystem)
+    network_cls = ecosystem_cls.get_network(network)
     with network_cls.use_provider("infura") as provider:
         yield provider
